feat: move updater to channel module
